feat: add conversion to u128; modify wad to ray conversion (#40)
This commit implements `Into<Wad, u128>` and `Into<Ray, u128>`. 

It also removes `TryInto<Wad, Ray>` in favour of a `wad_to_ray` 
helper function, which is necessary to prevent two different paths 
for converting wad to ray. This also better aligns with the scaling
that is done in this conversion, as opposed to a pure type conversion.